# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  domain.RegisterRequest:
    properties:
      email:
        type: string
      full_name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        minLength: 8
        type: string
      phone:
        type: string
    required:
    - email
    - full_name
    - password
    type: object
...
  domain.UpdateProfileRequest:
    properties:
      default_addr:
        maxLength: 500
        type: string
      full_name:
        maxLength: 100
        minLength: 2
        type: string
      phone:
        type: string
    type: object
...
  server.errorResponse:
    properties:
      error:
        type: string
    type: object
  server.successResponse:
    properties:
      data: {}
    type: object
host: localhost:8080
info:
  contact:
    email: support@bookstore.local
    name: API Support
  description: Polyglot persistence bookstore backend — PostgreSQL · MongoDB · Neo4j
    · Redis.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: HCMUS Bookstore API
  version: "2.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate and receive a JWT access token
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.successResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.errorResponse'
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      description: Revoke the current JWT token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.successResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new customer account
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/server.successResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/server.errorResponse'
      summary: Register
      tags:
      - auth
...
  /users/me:
    get:
      description: Return the current user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.successResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.errorResponse'
      security:
      - BearerAuth: []
      summary: Get profile
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update the current user's profile fields
      parameters:
      - description: Profile update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.successResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.errorResponse'
      security:
      - BearerAuth: []
      summary: Update profile
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"